scheduler: reduce size of relative_time_ms
We don't use a period longer than 2 seconds and reducing the relative
time type allows us to save 10 bytes of RAM and 268 bytes of flash
space.

Can't pass on these savings in this economy...

Before:
RAM:   [========  ]  81.5% (used 1670 bytes from 2048 bytes)
Flash: [========= ]  90.2% (used 29100 bytes from 32256 bytes)

After:
RAM:   [========  ]  81.1% (used 1660 bytes from 2048 bytes)
Flash: [========= ]  89.4% (used 28832 bytes from 32256 bytes)

+++ lib/scheduling/time.hpp
@@ -11,7 +11,7 @@
 namespace nsec::scheduling {
 
 using absolute_time_ms = unsigned long;
-using relative_time_ms = unsigned long;
+using relative_time_ms = unsigned short;
 
 } // namespace nsec::scheduling
